requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.9
Faker==36.1.1
inflection==0.5.1
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
rest-framework-simplejwt==0.0.2
simplejwt==2.0.1
sqlparse==0.5.3
typing==3.7.4.3
tzdata==2025.1
uritemplate==4.1.1
